+++ tests/ttnn/unit_tests/operations/eltwise/test_binary_uint32.py
@@ -8,6 +8,31 @@ import ttnn
 
 
 @pytest.mark.parametrize(
+    "ttnn_op, value_ranges",
+    [
+        (
+            ttnn.add,
+            [
+                (0, 100, 0, 300),
+                (1000, 10000, 500, 1000),
+                (30000, 40000, 10000, 15000),
+                (50000, 55000, 1000, 2000),
+                (0, 70000, 80000, 200000),
+            ],
+        ),
+        (
+            ttnn.sub,
+            [
+                (100, 500, 0, 100),
+                (1000, 10000, 500, 1000),
+                (30000, 40000, 10000, 15000),
+                (50000, 55000, 1000, 2000),
+                (80000, 200000, 0, 70000),
+            ],
+        ),
+    ],
+)
+@pytest.mark.parametrize(
     "a_shape, b_shape",
     [
         (torch.Size([1, 2, 32]), torch.Size([1, 2, 32])),
@@ -17,17 +42,11 @@ import ttnn
         (torch.Size([5]), torch.Size([1])),
     ],
 )
-@pytest.mark.parametrize(
-    "low_a, high_a, low_b, high_b",
-    [
-        (0, 100, 0, 300),
-        (1000, 10000, 500, 1000),
-        (30000, 40000, 10000, 15000),
-        (50000, 55000, 1000, 2000),
-        (0, 70000, 80000, 200000),
-    ],
-)
-def test_binary_add_uint32_bcast(a_shape, b_shape, low_a, high_a, low_b, high_b, device):
+@pytest.mark.parametrize("range_idx", [0, 1, 2, 3, 4])
+def test_binary_uint32_bcast(ttnn_op, value_ranges, a_shape, b_shape, range_idx, device):
+    """Test uint32 addition and subtraction with broadcast"""
+    low_a, high_a, low_b, high_b = value_ranges[range_idx]
+
     num_elements = max(int(torch.prod(torch.tensor(a_shape)).item()), 1)
     torch_input_tensor_a = torch.linspace(high_a, low_a, num_elements, dtype=torch.int32)
     torch_input_tensor_a = torch_input_tensor_a[:num_elements].reshape(a_shape).nan_to_num(0.0)
@@ -36,7 +55,7 @@ def test_binary_add_uint32_bcast(a_shape, b_shape, low_a, high_a, low_b, high_b,
     torch_input_tensor_b = torch.linspace(high_b, low_b, num_elements, dtype=torch.int32)
     torch_input_tensor_b = torch_input_tensor_b[:num_elements].reshape(b_shape).nan_to_num(0.0)
 
-    golden_function = ttnn.get_golden_function(ttnn.add)
+    golden_function = ttnn.get_golden_function(ttnn_op)
     torch_output_tensor = golden_function(torch_input_tensor_a, torch_input_tensor_b, device=device)
 
     input_tensor_a = ttnn.from_torch(
@@ -54,7 +73,8 @@ def test_binary_add_uint32_bcast(a_shape, b_shape, low_a, high_a, low_b, high_b,
         layout=ttnn.TILE_LAYOUT,
         memory_config=ttnn.DRAM_MEMORY_CONFIG,
     )
-    output_tensor = ttnn.add(input_tensor_a, input_tensor_b, use_legacy=None)
+
+    output_tensor = ttnn_op(input_tensor_a, input_tensor_b)
     output_tensor = ttnn.to_torch(output_tensor, dtype=torch.int32)
 
     assert torch.equal(output_tensor, torch_output_tensor)
@@ -86,6 +106,13 @@ block_sharded_memory_config = ttnn.create_sharded_memory_config(
 
 
 @pytest.mark.parametrize(
+    "ttnn_op, low_a, high_a, low_b, high_b",
+    [
+        (ttnn.add, 0, 100, 100, 200),  # Addition: ranges don't overlap to avoid overflow
+        (ttnn.sub, 50000, 100000, 0, 40000),  # Subtraction: ensure a > b for valid results
+    ],
+)
+@pytest.mark.parametrize(
     "a_shape, b_shape",
     ((torch.Size([5, 7, 64, 128]), torch.Size([5, 7, 64, 128])),),
 )
@@ -97,13 +124,14 @@ block_sharded_memory_config = ttnn.create_sharded_memory_config(
         block_sharded_memory_config,
     ],
 )
-def test_binary_add_uint32_sharded(a_shape, b_shape, sharded_config, device):
+def test_binary_uint32_sharded(ttnn_op, low_a, high_a, low_b, high_b, a_shape, b_shape, sharded_config, device):
+    """Test uint32 addition and subtraction with sharded memory configurations"""
     num_elements = max(int(torch.prod(torch.tensor(a_shape)).item()), 1)
-    torch_input_tensor_a = torch.linspace(0, 100, num_elements, dtype=torch.int32)
+    torch_input_tensor_a = torch.linspace(high_a, low_a, num_elements, dtype=torch.int32)
     torch_input_tensor_a = torch_input_tensor_a[:num_elements].reshape(a_shape).nan_to_num(0.0)
 
     num_elements = max(int(torch.prod(torch.tensor(b_shape)).item()), 1)
-    torch_input_tensor_b = torch.linspace(100, 200, num_elements, dtype=torch.int32)
+    torch_input_tensor_b = torch.linspace(high_b, low_b, num_elements, dtype=torch.int32)
     torch_input_tensor_b = torch_input_tensor_b[:num_elements].reshape(b_shape).nan_to_num(0.0)
 
     input_tensor_a = ttnn.from_torch(
@@ -121,13 +149,109 @@ def test_binary_add_uint32_sharded(a_shape, b_shape, sharded_config, device):
         memory_config=sharded_config,
     )
 
-    golden_function = ttnn.get_golden_function(ttnn.add)
+    golden_function = ttnn.get_golden_function(ttnn_op)
     torch_output_tensor = golden_function(torch_input_tensor_a, torch_input_tensor_b, device=device)
 
-    output_tensor = ttnn.add(input_tensor_a, input_tensor_b, memory_config=sharded_config, use_legacy=None)
+    output_tensor = ttnn_op(input_tensor_a, input_tensor_b, memory_config=sharded_config)
     output_tensor = ttnn.to_torch(output_tensor, dtype=torch.int32)
 
     assert torch.equal(output_tensor, torch_output_tensor)
+
+
+@pytest.mark.parametrize(
+    "a_shape, b_shape",
+    [
+        (torch.Size([1, 2, 32, 64, 125]), torch.Size([1, 2, 32, 1, 1])),
+    ],
+)
+@pytest.mark.parametrize(
+    "low_a, high_a, low_b, high_b",
+    [
+        (0, 100, 200, 300),  # Test underflow cases (a < b)
+        (500, 1000, 1000, 10000),
+        (10000, 15000, 30000, 40000),
+        (1000, 2000, 50000, 55000),
+        (0, 70000, 80000, 200000),
+    ],
+)
+def test_binary_sub_uint32_underflow(a_shape, b_shape, low_a, high_a, low_b, high_b, device):
+    """Test uint32 subtraction with underflow cases (wrap-around behavior)"""
+    num_elements = max(int(torch.prod(torch.tensor(a_shape)).item()), 1)
+    torch_input_tensor_a = torch.linspace(high_a, low_a, num_elements, dtype=torch.int32)
+    torch_input_tensor_a = torch_input_tensor_a[:num_elements].reshape(a_shape).nan_to_num(0.0)
+
+    num_elements = max(int(torch.prod(torch.tensor(b_shape)).item()), 1)
+    torch_input_tensor_b = torch.linspace(high_b, low_b, num_elements, dtype=torch.int32)
+    torch_input_tensor_b = torch_input_tensor_b[:num_elements].reshape(b_shape).nan_to_num(0.0)
+
+    golden_function = ttnn.get_golden_function(ttnn.sub)
+    torch_output_tensor = golden_function(torch_input_tensor_a, torch_input_tensor_b, device=device)
+
+    input_tensor_a = ttnn.from_torch(
+        torch_input_tensor_a,
+        dtype=ttnn.uint32,
+        device=device,
+        layout=ttnn.TILE_LAYOUT,
+        memory_config=ttnn.DRAM_MEMORY_CONFIG,
+    )
+
+    input_tensor_b = ttnn.from_torch(
+        torch_input_tensor_b,
+        dtype=ttnn.uint32,
+        device=device,
+        layout=ttnn.TILE_LAYOUT,
+        memory_config=ttnn.DRAM_MEMORY_CONFIG,
+    )
+    output_tensor = ttnn.sub(input_tensor_a, input_tensor_b)
+    output_tensor = ttnn.to_torch(output_tensor, dtype=torch.int32)
+
+    assert torch.equal(output_tensor, torch_output_tensor)
+
+
+@pytest.mark.parametrize("use_legacy", [True, False])
+def test_binary_sub_uint32_edge_cases(use_legacy, device):
+    """Test uint32 subtraction with edge cases including underflow"""
+    torch_input_tensor_a = torch.tensor(
+        [4294967295, 2147483647, 1000000000, 500, 0, 100, 1, 0]  # uint32 max, int32 max, large, medium, min
+    )
+    input_tensor_a = ttnn.from_torch(
+        torch_input_tensor_a,
+        dtype=ttnn.uint32,
+        device=device,
+        layout=ttnn.TILE_LAYOUT,
+        memory_config=ttnn.DRAM_MEMORY_CONFIG,
+    )
+    torch_input_tensor_b = torch.tensor(
+        [1, 1000000000, 2147483647, 1000, 1, 50, 2, 1]  # Various values to test different scenarios
+    )
+    input_tensor_b = ttnn.from_torch(
+        torch_input_tensor_b,
+        dtype=ttnn.uint32,
+        device=device,
+        layout=ttnn.TILE_LAYOUT,
+        memory_config=ttnn.DRAM_MEMORY_CONFIG,
+    )
+
+    # Expected output values for uint32 subtraction with these specific inputs
+    expected_output_values = torch.tensor(
+        [4294967294, 1147483647, 3147483649, 4294966796, 4294967295, 50, 4294967295, 4294967295], dtype=torch.uint32
+    )
+    expected_tensor = ttnn.from_torch(
+        expected_output_values,
+        dtype=ttnn.uint32,
+        device=device,
+        layout=ttnn.TILE_LAYOUT,
+        memory_config=ttnn.DRAM_MEMORY_CONFIG,
+    )
+
+    output_tensor = ttnn.sub(input_tensor_a, input_tensor_b, use_legacy=use_legacy)
+
+    # Compare using ttnn.eq and check if all results are ones (all elements match)
+    comparison_result = ttnn.eq(output_tensor, expected_tensor)
+    comparison_torch = ttnn.to_torch(comparison_result, dtype=torch.bool)
+
+    # Verify all comparisons are True (all elements match)
+    assert torch.all(comparison_torch), f"Mismatch found in uint32 subtraction results"
 
 
 # For inputs within int32 range [0, 2147483647]
